+++ R/imports.R
@@ -5,6 +5,7 @@
 ## #' @importFrom taxize tnrs
 #' @importFrom foreign read.dbf
 #' @importFrom stringdist stringsim
+#' @importFrom stringr fixed str_replace_all str_split
 #' @importClassesFrom biblio lib_df
 # #' @importFrom taxa taxon_ids taxon_ranks
 # #' @importClassesFrom biblio lib_df